backend/pyproject.toml
[project]
name = "backend-modular"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "deepface>=0.0.93",
    "deepgram-sdk==3.2.6",
    "fastapi>=0.120.0",
    "google-api-python-client>=2.185.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pylast>=6.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sounddevice>=0.5.3",
    "tensorflow-cpu>=2.16.0",
    "tf-keras>=2.20.1",
    "uvicorn[standard]>=0.38.0",
]
